requirements.txt
python-dotenv==1.0.1
python-telegram-bot==13.7
standard-imghdr==3.13.0
urllib3==1.26.15
